requirements.txt
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
distlib==0.3.7
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-redis==5.4.0
django-timezone-field==6.1.0
docopt==0.6.2
flake8==6.1.0
fontawesomefree==6.4.2
humanize==4.8.0
idna==3.6
isort==5.13.0
kombu==5.3.4
mccabe==0.7.0
pep517==0.13.1
Pillow==10.0.1
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
prompt-toolkit==3.0.41
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
requirementslib==3.0.0
six==1.16.0
sqlparse==0.4.4
statuses==0.5.2
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
yarg==0.1.9
